fix(storybook): move stories under src
Having them out of src is breaking tslint and some other processes.

BREAKING CHANGE: `/stories` folder must be moved to `/src/__stories`, and `!__stories__` must be
added to the package.json `files` property.
