+++ .storybook/config.js
@@ -37,7 +37,7 @@ addDecorator(
 addDecorator(withThemes(themes));
 
 function loadStories() {
-  require('../../../../stories');
+  require('../../../../src/__stories__');
 }
 
 configure(loadStories, module);
