+++ templates/common/tsconfig.json
@@ -1,5 +1,6 @@
 {
   "extends": "./node_modules/@stoplight/scripts/tsconfig.json",
+  // NOTE: must only include one element, otherwise build process into dist folder ends up with incorrect structure
   "include": ["src"],
   "compilerOptions": {
     "outDir": "dist"
