+++ templates/pkgs/base.json
@@ -13,7 +13,7 @@
   },
   "license": "UNLICENSED",
   "main": "src/index.ts",
-  "files": ["**/*", "!__tests__"],
+  "files": ["**/*", "!__tests__", "!__stories__"],
   "engines": {
     "node": ">=8.3.0"
   },
