+++ templates/srcs/react/src/__stories__/Button.tsx
@@ -3,7 +3,7 @@ import * as React from 'react';
 import { action } from '@storybook/addon-actions';
 import { storiesOf } from '@storybook/react';
 
-import { Button } from '../src/Button';
+import { Button } from '../Button';
 
 storiesOf('Button', module)
   // @ts-ignore
