+++ templates/srcs/react/src/__stories__/Welcome.tsx
