+++ templates/srcs/react/src/__stories__/index.ts
